feat: FileList.jsx에서 FileComponent.jsx를 분리
